requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
aiofiles
httpx
python-dotenv
litellm
websockets
pydantic-settings
